完善

+++ framework_api/src/main/java/com/trkj/framework/entity/mybatisplus/DeptPost.java
@@ -26,14 +26,14 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 @TableName("DEPT_POST")
 @ApiModel(value="DeptPost对象", description="部门职位表")
-@KeySequence(value = "DEPT_POST_ID",clazz = Integer.class)
+@KeySequence(value = "DEPT_POST_ID",clazz = Long.class)
 public class DeptPost implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
     @ApiModelProperty(value = "部门职位编号")
     @TableId("DEPT_POST_ID")
-    private Integer deptPostId;
+    private Long deptPostId;
 
     @ApiModelProperty(value = "部门编号")
     @TableField("DEPT_ID")
